Auto Rotate by Exif Orientation
Adds the functionality to rotate the images (detail, thumb, original) by the orientation in the exif metadata. If there are no such metadata available nothing will happen. There are two ways of rotating the image if possible it will happen with imagemagick (it needs much less computing power and ram). If it is not available on the server the rotation will be done by GD.