Additional Fields to Frontend Form
I developed JoomGallery further, to finally make it possible to easily have the Additional Image Fields ready to use directly in the frontend forms when uploading images.

( Note : You'll still need to additionally install the Plugin "JoomAdditionalImageFields" for you to make use of that feature in your JoomGallery installation. If you don't, it doesnt affect you.
Note2: I also updated the "JoomAdditionalImageFields" accordingly - You'll need to use that new version together with this update, if you want the Additional Image Fileds be available in frontend forms ).